requirements.txt
asgiref==3.6.0
autopep8==2.0.1
boto3==1.26.107
botocore==1.29.107
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==39.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
django-rest-knox==4.2.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dynaconf==3.1.12
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
hvac==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
protobuf==4.22.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydotenv==0.0.7
pyhcl==0.4.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
